news/serializers.py
from .models import (
    Post,
    Category,
)
from rest_framework import serializers


class PostSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Post
        fields = [
            'id',
            'author',
            'title',
            'time_pub',
            'rating',
        ]


class CategorySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Category
        fields = ['id', 'cat_name', ]
