requirements.txt
amqp==5.1.1
APScheduler==3.9.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==37.0.2
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-allauth==0.50.0
django-apscheduler==0.6.2
django-filter==21.1
django-modeltranslation==0.18.3
idna==3.3
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.29
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1

djangorestframework~=3.13.1